refactor: LottoWinningRankingCalculator를 LottoWinningRankingCountService로 변경
로또 당첨 결과 개수를 세는 기능은 로또에 대한 "도메인"이 아닌 "서비스"다.
